+++ src/containers/InteractionMonitoring/layout.js
@@ -22,7 +22,6 @@ import elapsedTimeColumn from '../../components/tableColumns/elapsedTime';
 import monitoringColumn from '../../components/tableColumns/monitoring';
 import groupsColumn from '../../components/tableColumns/groups';
 import skillsColumn from '../../components/tableColumns/skills';
-import { sdkCall } from '../../utils/sdk';
 
 const InlineCheckboxFilterMenu = styled(CheckboxFilterMenu)`
   display: inline-block;
@@ -64,12 +63,6 @@ export default class InteractionMonitoring extends Component {
     this.props.fetchData('groups', 'interactionMonitoring');
     this.props.fetchData('skills', 'interactionMonitoring');
     this.props.startInteractionMonitoring();
-  }
-
-  componentDidMount() {
-    if (this.props.insideIframe) {
-      sdkCall({ module: 'removeDirtyFormIdFromSessionStorage' });
-    }
   }
 
   componentWillUnmount() {